fix: moved CategoriesCheckboxGroup component to the features folder
