+++ src/features/categories/types/index.ts
@@ -1,0 +1,6 @@
+export interface ICategoryForm {
+  icon: string;
+  text: string;
+  value: number;
+  isChecked: boolean;
+}
